+++ packages/layout/src/container.tsx
@@ -1,21 +1,15 @@
 import {
   chakra,
   forwardRef,
-  SystemStyleObject,
-  useTheme,
+  omitThemingProps,
+  ThemingProps,
+  useStyleConfig,
   HTMLChakraProps,
 } from "@chakra-ui/system"
-import {
-  cx,
-  Dict,
-  filterUndefined,
-  mapResponsive,
-  memoizedGet as get,
-  __DEV__,
-} from "@chakra-ui/utils"
+import { cx, __DEV__ } from "@chakra-ui/utils"
 import * as React from "react"
 
-export interface ContainerProps extends HTMLChakraProps<"div"> {
+export interface ContainerProps extends HTMLChakraProps<"div">, ThemingProps {
   /**
    * If `true`, container will center its children
    * regardless of their width.
@@ -35,65 +29,27 @@ export const Container = forwardRef<ContainerProps, "div">(function Container(
   props,
   ref,
 ) {
-  const {
-    maxWidth,
-    width,
-    minWidth,
-    w,
-    minW,
-    maxW,
-    className,
-    centerContent,
-    ...rest
-  } = props
+  const { className, centerContent, ...rest } = omitThemingProps(props)
 
-  const theme = useTheme()
-
-  const widthProps = transform(theme, {
-    maxW,
-    maxWidth,
-    width,
-    w,
-    minWidth,
-    minW,
-  })
-
-  const styles: SystemStyleObject = {
-    w: "100%",
-    mx: "auto",
-    maxW: "60ch",
-    px: "1rem",
-    ...(centerContent && {
-      display: "flex",
-      flexDirection: "column",
-      alignItems: "center",
-    }),
-    ...widthProps,
-  }
+  const styles = useStyleConfig("Container", props)
 
   return (
     <chakra.div
       ref={ref}
       className={cx("chakra-container", className)}
       {...rest}
-      __css={styles}
+      __css={{
+        ...styles,
+        ...(centerContent && {
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "center",
+        }),
+      }}
     />
   )
 })
 
 if (__DEV__) {
   Container.displayName = "Container"
-}
-
-function transform(theme: Dict, props: Dict) {
-  const result: SystemStyleObject = {}
-
-  Object.keys(props).forEach((prop) => {
-    const propValue = props[prop]
-    result[prop] = mapResponsive(propValue, (value) =>
-      get(theme, `sizes.container.${value}`, value),
-    )
-  })
-
-  return filterUndefined(result)
 }