+++ packages/layout/src/stack.tsx
@@ -3,16 +3,14 @@ import {
   ChakraComponent,
   forwardRef,
   HTMLChakraProps,
-  PropsOf,
-  ResponsiveValue,
   SystemProps,
 } from "@chakra-ui/system"
-import { cx, getValidChildren, mapResponsive, __DEV__ } from "@chakra-ui/utils"
+import { cx, getValidChildren, __DEV__ } from "@chakra-ui/utils"
 import * as React from "react"
+import type { StackDirection } from "./stack.utils"
+import { getDividerStyles, getStackStyles, selector } from "./stack.utils"
 
-export type StackDirection = ResponsiveValue<
-  "row" | "column" | "row-reverse" | "column-reverse"
->
+export type { StackDirection }
 
 interface StackOptions {
   /**
@@ -63,7 +61,7 @@ export const StackDivider: ChakraComponent<"div"> = (props) => {
       className="chakra-stack__divider"
       {...props}
       __css={{
-        ...props.__css,
+        ...props["__css"],
         borderWidth: 0,
         alignSelf: "stretch",
         borderColor: "inherit",
@@ -87,13 +85,7 @@ export const StackItem: ChakraComponent<"div"> = (props) => (
   />
 )
 
-export interface StackProps extends PropsOf<typeof chakra.div>, StackOptions {}
-
-/**
- * If we ever run into SSR issues with this, check this post to find a fix for it:
- * @see https://medium.com/@emmenko/patching-lobotomized-owl-selector-for-emotion-ssr-5a582a3c424c
- */
-const selector = "& > *:not(style) ~ *:not(style)"
+export interface StackProps extends HTMLChakraProps<"div">, StackOptions {}
 
 /**
  * Stacks help you easily create flexible and automatically distributed layouts
@@ -123,45 +115,15 @@ export const Stack = forwardRef<StackProps, "div">(function Stack(props, ref) {
 
   const direction = isInline ? "row" : directionProp ?? "column"
 
-  const styles = React.useMemo(() => {
-    const directionStyles = {
-      column: { mt: spacing, ml: 0 },
-      row: { ml: spacing, mt: 0 },
-      "column-reverse": { mb: spacing, mr: 0 },
-      "row-reverse": { mr: spacing, mb: 0 },
-    }
+  const styles = React.useMemo(() => getStackStyles({ direction, spacing }), [
+    direction,
+    spacing,
+  ])
 
-    return {
-      flexDirection: direction,
-      [selector]: mapResponsive(direction, (value) => directionStyles[value]),
-    }
-  }, [direction, spacing])
-
-  /**
-   * Divider Styles
-   */
-  const marginX = mapResponsive(direction, (value) =>
-    value.includes("row") ? spacing : 0,
+  const dividerStyle = React.useMemo(
+    () => getDividerStyles({ spacing, direction }),
+    [spacing, direction],
   )
-
-  const marginY = mapResponsive(direction, (value) =>
-    value.includes("row") ? 0 : spacing,
-  )
-
-  const borderLeftWidth = mapResponsive(direction, (value) =>
-    value.includes("row") ? "1px" : 0,
-  )
-
-  const borderBottomWidth = mapResponsive(direction, (value) =>
-    value.includes("row") ? 0 : "1px",
-  )
-
-  const dividerStyles = {
-    marginX,
-    marginY,
-    borderLeftWidth,
-    borderBottomWidth,
-  }
 
   const hasDivider = !!divider
   const shouldUseChildren = !shouldWrapChildren && !hasDivider
@@ -177,10 +139,19 @@ export const Stack = forwardRef<StackProps, "div">(function Stack(props, ref) {
 
         if (!hasDivider) return _child
 
-        const clonedDivider = React.cloneElement(divider as any, dividerStyles)
+        const clonedDivider = React.cloneElement(
+          divider as React.ReactElement<any>,
+          { __css: dividerStyle },
+        )
+
         const _divider = isLast ? null : clonedDivider
 
-        return <React.Fragment key={index}>{[_child, _divider]}</React.Fragment>
+        return (
+          <React.Fragment key={index}>
+            {_child}
+            {_divider}
+          </React.Fragment>
+        )
       })
 
   const _className = cx("chakra-stack", className)