+++ packages/progress/stories/progress.stories.tsx
@@ -1,6 +1,6 @@
+import { chakra } from "@chakra-ui/system"
 import * as React from "react"
 import { Progress, ProgressLabel } from "../src"
-import { chakra } from "@chakra-ui/system"
 
 export default {
   title: "Linear Progress",
@@ -43,4 +43,8 @@ export const withSizes = () => (
 
 export const withAnimation = () => (
   <Progress colorScheme="green" hasStripe isAnimated value={20} />
+)
+
+export const withCustomBorderRadius = () => (
+  <Progress value={20} borderRadius="4px" />
 )