+++ packages/theme/src/components/divider.ts
@@ -1,0 +1,27 @@
+const baseStyle = {
+  opacity: 0.6,
+  borderColor: "inherit",
+}
+
+const variantSolid = {
+  borderStyle: "solid",
+}
+
+const variantDashed = {
+  borderStyle: "dashed",
+}
+
+const variants = {
+  solid: variantSolid,
+  dashed: variantDashed,
+}
+
+const defaultProps = {
+  variant: "solid",
+}
+
+export default {
+  baseStyle,
+  variants,
+  defaultProps,
+}