+++ packages/theme/src/components/heading.ts
@@ -14,19 +14,19 @@ const sizes = {
   },
   "2xl": {
     fontSize: ["4xl", null, "5xl"],
-    lineHeight: ["2.5rem", null, "1"],
+    lineHeight: [1.2, null, 1],
   },
   xl: {
     fontSize: ["3xl", null, "4xl"],
-    lineHeight: ["2.25rem", null, "2.5rem"],
+    lineHeight: [1.33, null, 1.2],
   },
   lg: {
     fontSize: ["2xl", null, "3xl"],
-    lineHeight: ["2rem", null, "2.25rem"],
+    lineHeight: [1.33, null, 1.2],
   },
-  md: { fontSize: "xl", lineHeight: "1.75rem" },
-  sm: { fontSize: "md", lineHeight: "1.5rem" },
-  xs: { fontSize: "sm", lineHeight: "1.25rem" },
+  md: { fontSize: "xl", lineHeight: 1.2 },
+  sm: { fontSize: "md", lineHeight: 1.2 },
+  xs: { fontSize: "sm", lineHeight: 1.2 },
 }
 
 const defaultProps = {