+++ website/pages/index.tsx
@@ -126,9 +126,12 @@ const HomePage = ({ members, sponsors }: HomePageProps) => {
           pb={{ base: "0", md: "5rem" }}
         >
           <Container>
-            <Box maxW="760px" mx="auto" textAlign="center">
+            <Box textAlign="center">
               <chakra.h1
-                fontSize={{ base: "2.25rem", sm: "3rem", lg: "3.75rem" }}
+                maxW="680px"
+                mx="auto"
+                fontSize={{ base: "2.25rem", sm: "3rem", lg: "4rem" }}
+                fontFamily="heading"
                 letterSpacing="tight"
                 fontWeight="bold"
                 mb="16px"
@@ -144,7 +147,13 @@ const HomePage = ({ members, sponsors }: HomePageProps) => {
                 </Box>
               </chakra.h1>
 
-              <Text opacity={0.7} fontSize={{ base: "lg", lg: "xl" }} mt="6">
+              <Text
+                maxW="560px"
+                mx="auto"
+                opacity={0.7}
+                fontSize={{ base: "lg", lg: "xl" }}
+                mt="6"
+              >
                 Chakra UI is a simple, modular and accessible component library
                 that gives you the building blocks you need to build your React
                 applications.
@@ -194,7 +203,7 @@ const HomePage = ({ members, sponsors }: HomePageProps) => {
                 px="6"
                 py="4"
               >
-                <Img h="64px" src="/git-nation-badge.png" />
+                <Img h="55px" src="/git-nation-badge.png" />
               </Box>
             </Center>
           </Container>