+++ website/src/components/algolia-search.tsx
@@ -1,142 +1,181 @@
-import {
-  Box,
-  Input,
-  InputGroup,
-  InputLeftElement,
-  useEventListener,
-} from "@chakra-ui/react"
 import { SearchIcon } from "@chakra-ui/icons"
-import { get, startsWith } from "lodash/fp"
+import {
+  chakra,
+  HStack,
+  HTMLChakraProps,
+  Kbd,
+  Portal,
+  Text,
+  useColorModeValue,
+  VisuallyHidden,
+} from "@chakra-ui/react"
+import { DocSearchModal, useDocSearchKeyboardEvents } from "@docsearch/react"
+import { startsWith } from "lodash/fp"
+import Head from "next/head"
+import Link from "next/link"
 import { useRouter } from "next/router"
-import React from "react"
+import * as React from "react"
+import SearchStyle from "./search.styles"
 
-const getLvl1 = get("hierarchy.lvl1")
 const startsWithCss = startsWith("css-")
 
-let link: HTMLAnchorElement
+const ACTION_KEY_DEFAULT = ["Ctrl", "Control"]
+const ACTION_KEY_APPLE = ["⌘", "Command"]
 
-if (typeof window !== "undefined") {
-  link = document.createElement("a")
-}
-
-function getHash(url: string) {
-  link.href = url
-  return link.hash
-}
-
-function Search(props) {
-  const router = useRouter()
-  const ref = React.useRef<HTMLInputElement>()
-
-  const onKeyDown = (event: KeyboardEvent) => {
-    if (event.key === "/") {
-      const activeElement = document.activeElement
-      const focusWrapper = document.getElementById("gatsby-focus-wrapper")
-
-      if (activeElement === focusWrapper || activeElement === document.body) {
-        event.preventDefault()
-        ref.current?.focus({ preventScroll: true })
-      }
-    }
-  }
-
-  useEventListener("keydown", onKeyDown)
-
-  React.useEffect(() => {
-    if (window) {
-      import("docsearch.js").then(({ default: docsearch }) => {
-        const _window = window as any
-        _window.docsearch = docsearch
-        docsearch({
-          apiKey: "df1dcc41f7b8e5d68e73dd56d1e19701",
-          indexName: "chakra-ui",
-          inputSelector: "#algolia-search",
-          // debug: true,
-          handleSelected: (input, event, suggestion) => {
-            event.preventDefault()
-            input.setVal("")
-            input.close()
-            if (ref.current) {
-              ref.current.blur()
-            }
-
-            const url = suggestion.url.replace("https://chakra-ui.com", "")
-            router.push(url)
-            const hash = window.decodeURI(getHash(url))
-
-            if (hash !== "#" && hash !== "") {
-              const link: HTMLAnchorElement = document.querySelector(
-                `.docSearch-content ${hash} a`,
-              )
-              if (link) {
-                link.click()
-              }
-            }
-          },
-          transformData(hits: any[]) {
-            return hits.filter((hit) => {
-              const lvl1 = getLvl1(hit)
-              return !startsWithCss(lvl1)
-            })
-          },
-        })
-      })
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
+function Hit(props) {
+  const { hit, children } = props as any
   return (
-    <Box
-      position="relative"
-      width="100%"
-      mt="3"
-      mb="32px"
-      boxSizing="content-box"
-      display={["none", "block", "block"]}
-      {...props}
-      sx={{
-        ".algolia-autocomplete": {
-          width: "100%",
-        },
-        ".algolia-autocomplete .ds-dropdown-menu": {
-          width: "100% !important",
-          maxWidth: "100% !important",
-          minWidth: "0 !important",
-        },
-        ".algolia-docsearch-suggestion--category-header": {
-          bg: "teal.400",
-        },
-        ".algolia-docsearch-suggestion--highlight": {
-          bg: "teal.50",
-          color: "gray.800",
-        },
-        ".ds-dropdown-menu": {
-          "&:before": {
-            display: "none",
-          },
-        },
-        ".ds-cursor .algolia-docsearch-suggestion--wrapper": {
-          bg: "gray.100",
-          boxShadow: "none",
-        },
-      }}
-    >
-      <InputGroup size="md">
-        <InputLeftElement>
-          <SearchIcon color="gray.500" />
-        </InputLeftElement>
-        <Input
-          bg="gray.50"
-          placeholder={`Search docs`}
-          focusBorderColor="teal.200"
-          ref={ref}
-          borderRadius="lg"
-          id="algolia-search"
-          aria-label="Search Chakra UI docs"
-        />
-      </InputGroup>
-    </Box>
+    <Link href={hit.url}>
+      <a>{children}</a>
+    </Link>
   )
 }
 
-export default Search
+export const SearchButton = React.forwardRef(function SearchButton(
+  props: HTMLChakraProps<"button">,
+  ref: React.Ref<HTMLButtonElement>,
+) {
+  const [actionKey, setActionKey] = React.useState<string[]>(ACTION_KEY_APPLE)
+  React.useEffect(() => {
+    if (typeof navigator === "undefined") return
+    const isMac = /(Mac|iPhone|iPod|iPad)/i.test(navigator.platform)
+    if (!isMac) {
+      setActionKey(ACTION_KEY_DEFAULT)
+    }
+  }, [])
+
+  return (
+    <chakra.button
+      flex="1"
+      type="button"
+      mx="6"
+      ref={ref}
+      lineHeight="1.2"
+      w="100%"
+      bg={useColorModeValue("white", "gray.700")}
+      whiteSpace="nowrap"
+      display={{ base: "none", sm: "flex" }}
+      alignItems="center"
+      color="gray.400"
+      py="3"
+      px="4"
+      outline="0"
+      _focus={{ shadow: "outline" }}
+      shadow="base"
+      rounded="md"
+      {...props}
+    >
+      <SearchIcon />
+      <HStack w="full" ml="3" spacing="4px">
+        <Text textAlign="left" flex="1">
+          Search the docs
+        </Text>
+        <HStack spacing="4px">
+          <VisuallyHidden>Press </VisuallyHidden>
+          <Kbd color="gray.500" rounded="2px">
+            <chakra.div
+              as="abbr"
+              title={actionKey[1]}
+              textDecoration="none !important"
+            >
+              {ACTION_KEY_APPLE[0]}
+            </chakra.div>
+          </Kbd>
+          <VisuallyHidden> and </VisuallyHidden>
+          <Kbd color="gray.500" rounded="2px">
+            K
+          </Kbd>
+          <VisuallyHidden> to search</VisuallyHidden>
+        </HStack>
+      </HStack>
+    </chakra.button>
+  )
+})
+
+export function Search() {
+  const router = useRouter()
+  const [isOpen, setIsOpen] = React.useState(false)
+  const searchButtonRef = React.useRef()
+  const [initialQuery, setInitialQuery] = React.useState(null)
+
+  const onOpen = React.useCallback(() => {
+    setIsOpen(true)
+  }, [setIsOpen])
+
+  const onClose = React.useCallback(() => {
+    setIsOpen(false)
+  }, [setIsOpen])
+
+  const onInput = React.useCallback(
+    (e) => {
+      setIsOpen(true)
+      setInitialQuery(e.key)
+    },
+    [setIsOpen, setInitialQuery],
+  )
+
+  useDocSearchKeyboardEvents({
+    isOpen,
+    onOpen,
+    onClose,
+    onInput,
+    searchButtonRef,
+  })
+
+  return (
+    <>
+      <Head>
+        <link
+          rel="preconnect"
+          href="https://BH4D9OD16A-dsn.algolia.net"
+          crossOrigin="true"
+        />
+      </Head>
+      <SearchStyle />
+      <SearchButton onClick={onOpen} ref={searchButtonRef} />
+      {isOpen && (
+        <Portal>
+          <DocSearchModal
+            placeholder="Search the docs"
+            initialQuery={initialQuery}
+            initialScrollY={window.scrollY}
+            onClose={onClose}
+            indexName="chakra-ui"
+            apiKey="df1dcc41f7b8e5d68e73dd56d1e19701"
+            appId="BH4D9OD16A"
+            //@ts-expect-error
+            navigator={{
+              navigate({ suggestionUrl }) {
+                setIsOpen(false)
+                router.push(suggestionUrl)
+              },
+            }}
+            hitComponent={Hit}
+            transformItems={(items) => {
+              return items
+                .filter((item) => {
+                  const lvl1 = item.hierarchy.lvl1
+                  return !startsWithCss(lvl1) || !startsWithCss(item.content)
+                })
+                .map((item) => {
+                  /**
+                   *  We transform the absolute URL into a relative URL to
+                   *  leverage Next's preloading.
+                   */
+                  const a = document.createElement("a")
+                  a.href = item.url
+                  const hash = a.hash === "#content-wrapper" ? "" : a.hash
+
+                  return {
+                    ...item,
+                    url: `${a.pathname}${hash}`,
+                    content: item.content ?? item.hierarchy.lvl0,
+                  }
+                })
+            }}
+          />
+        </Portal>
+      )}
+    </>
+  )
+}