+++ packages/tabs/CHANGELOG.md
@@ -1,15 +1,31 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## 1.0.1
+
+### Patch Changes
+
+- [`892ea2ca`](https://github.com/chakra-ui/chakra-ui/commit/892ea2ca1c02b4127f4f044df33de58cc7641f5c)
+  Thanks [@segunadebayo](https://github.com/segunadebayo)! - - Fix orientation
+  styles for Tabs when orientation prop is vertical
+  [#1187](https://github.com/chakra-ui/chakra-ui/issues/1187)
+
+  - Add support for styling the `root` tab element from theme
+    [#2548](https://github.com/chakra-ui/chakra-ui/issues/2548)
+
+- Updated dependencies
+  [[`5c482483`](https://github.com/chakra-ui/chakra-ui/commit/5c482483ce24fc798540c9792a15e06772eae213),
+  [`204ff7e3`](https://github.com/chakra-ui/chakra-ui/commit/204ff7e39dd37003786194704b36069ef94d56a6)]:
+  - @chakra-ui/utils@1.0.1
+  - @chakra-ui/hooks@1.0.1
+  - @chakra-ui/clickable@1.0.1
+  - @chakra-ui/descendant@1.0.1
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 # 1.0.0 (2020-11-13)
 
 **Note:** Version bump only for package @chakra-ui/tabs
-
-
-
-
 
 # Change Log